--- modules/react-native-onboarding-v2/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Onboarding-v2 module",
-  "description": "Help your users get started with an onboarding tutorial welcome set of screens.",
-  "root": "/modules/onboarding-v2"
-}
