--- modules/react-native-onboarding-v2/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@modules/onboarding-v2",
-  "version": "1.0.0",
-  "description": "Help your users get started with an onboarding tutorial welcome set of screens.",
-  "private": true,
-  "main": "index.js",
-  "x-dependencies": {},
-  "author": "Crowdbotics",
-  "license": "ISC",
-  "dependencies": {
-    "react-native-app-intro-slider": "^4.0.4",
-    "prop-types" : "^15.8.1"
-    }
-}
