+++ modules/react-native-onboarding/meta.json
@@ -1,5 +1,5 @@
 {
-  "title": "Onboarding",
-  "description": "Help your users get started with an onboarding tutorial welcome set of screens.",
+  "title": "Onboarding module",
+  "description": "Help your users get started with an onboarding tutorial welcome set of screens. ",
   "root": "/modules/onboarding"
 }
