--- modules/react-native-onboarding/slides.js
@@ -1,29 +1,0 @@
-export const slides = [
-  {
-    key: "one",
-    title: "Title 1",
-    text: "Description.\nSay something cool",
-    image: {
-      uri:
-        "https://crowdbotics-slack-dev.s3.amazonaws.com/media/project_component_resources/cb-icon.png"
-    },
-    backgroundColor: "#59b2ab"
-  },
-  {
-    key: "two",
-    title: "Title 2",
-    text: "Other cool stuff",
-    image: {
-      uri:
-        "https://crowdbotics-slack-dev.s3.amazonaws.com/media/project_component_resources/cb-icon.png"
-    },
-    backgroundColor: "#febe29"
-  }
-  // {
-  //   key: 'three',
-  //   title: 'Rocket guy',
-  //   text: 'I\'m already out of descriptions\n\nLorem ipsum bla bla bla',
-  //   image: require('./assets/3.jpg'),
-  //   backgroundColor: '#22bcb5',
-  // }
-];
